requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.3
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2
django-allauth==0.61.1
django-bootstrap5==23.4
django-debug-toolbar==4.2.0
django-filter==23.5
django-mathfilters==1.0.0
django-taggit==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Faker==20.1.0
gunicorn==22.0.0
idna==3.6
inflection==0.5.1
Markdown==3.5.1
oauthlib==3.2.2
packaging==23.2
Pillow==10
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
stripe==9.4.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
